requirements.txt
aiofiles==0.6.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
Pillow==8.1.0
pydantic==1.7.3
python-multipart==0.0.5
pyzbar==0.1.8
qrcode==6.1
six==1.15.0
starkbank-brcode==1.0.1
starlette==0.13.6
uvicorn==0.13.3
